Remove last few hints of nerdhub packages
