+++ cardinal-components-util/src/main/java/dev/onyxstudios/cca/api/v3/util/Components.java
@@ -20,7 +20,7 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package nerdhub.cardinal.components.api.util;
+package dev.onyxstudios.cca.api.v3.util;
 
 import dev.onyxstudios.cca.api.v3.component.Component;
 import dev.onyxstudios.cca.api.v3.component.ComponentContainer;
@@ -78,5 +78,4 @@ public final class Components {
             op.accept(key, key.get(provider));
         }
     }
-
 }
