+++ cardinal-components-util/src/main/java/dev/onyxstudios/cca/api/v3/util/ObjectPath.java
@@ -20,7 +20,7 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package nerdhub.cardinal.components.api.util;
+package dev.onyxstudios.cca.api.v3.util;
 
 import javax.annotation.Nullable;
 import java.util.NoSuchElementException;
